Characters level 30 or over will now be hidden to the player on their character select list instead of actually being deleted when a player chooses to delete the character. Required SQL is handled by world automagically.
The is_deleted column in character_data is what hides the toon. Apps such as Magelo will need to be updated to handle this column, otherwise players will see their "hidden" characters.

Added #undeletechar command so GMs can restore these toons in-game.

+++ world/client.cpp
@@ -564,9 +564,17 @@ bool Client::HandleEnterWorldPacket(const EQApplicationPacket *app) {
 bool Client::HandleDeleteCharacterPacket(const EQApplicationPacket *app) {
 
 	uint32 char_acct_id = database.GetAccountIDByChar((char*)app->pBuffer);
+	uint32 level = database.GetLevelByChar((char*)app->pBuffer);
 	if(char_acct_id == GetAccountID()) {
 		Log.Out(Logs::Detail, Logs::World_Server,"Delete character: %s",app->pBuffer);
-		database.DeleteCharacter((char *)app->pBuffer);
+		if(level >= 30)
+		{
+			database.MarkCharacterDeleted((char *)app->pBuffer);
+		}
+		else
+		{
+			database.DeleteCharacter((char *)app->pBuffer);
+		}
 		SendCharInfo();
 	}
